requirement.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==4.2.16
django-filter==24.3
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
pillow==11.0.0
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.5.1
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
